django-backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psutil==5.9.6
torch>=2.0.0
transformers>=4.30.0
drf-spectacular>=0.26.5
pyyaml>=6.0

